requirements.txt
streamlit
pandas
numpy
scipy
scikit-learn
joblib
pyarrow
scikit-surprise
sentence-transformers
faiss-cpu
